+++ pyvirtualcam/camera.py
@@ -2,10 +2,11 @@ from typing import Optional
 import platform
 import time
 import warnings
+from enum import Enum
 
 import numpy as np
 
-from pyvirtualcam.util import FPSCounter
+from pyvirtualcam.util import FPSCounter, fourcc
 
 BACKENDS = {}
 
@@ -19,9 +20,21 @@ elif platform.system() == 'Linux':
     from pyvirtualcam import _native_linux_v4l2loopback
     BACKENDS['v4l2loopback'] = _native_linux_v4l2loopback.Camera
 
+class PixelFormat(Enum):
+    # See external/libyuv/include/libyuv/video_common.h for fourcc codes.
+    RGB = 'raw ', lambda w, h: (h, w, 3)
+    BGR = '24BG', lambda w, h: (h, w, 3)
+    GRAY = 'J400', lambda w, h: (h, w)
+    I420 = 'I420', lambda w, h: (w + w // 2) * h
+    YUYV = 'YUY2', lambda w, h: h * w * 2
+
 class Camera:
-    def __init__(self, width: int, height: int, fps: float, delay=None,
-                 print_fps=False, backend: Optional[str]=None, **kw) -> None:
+    def __init__(self, width: int, height: int, fps: float, *,
+                 fmt: PixelFormat=PixelFormat.RGB,
+                 backend: Optional[str]=None,
+                 print_fps=False,
+                 delay=None,
+                 **kw) -> None:
         if backend:
             backends = [(backend, BACKENDS[backend])]
         else:
@@ -30,7 +43,9 @@ class Camera:
         errors = []
         for name, clazz in backends:
             try:
-                self._backend = clazz(width=width, height=height, fps=fps, **kw)
+                self._backend = clazz(
+                    width=width, height=height, fps=fps, fmt=fourcc(fmt.value[0]),
+                    **kw)
             except Exception as e:
                 errors.append(f"'{name}' backend: {e}")
             else:
@@ -42,7 +57,20 @@ class Camera:
         self._width = width
         self._height = height
         self._fps = fps
+        self._fmt = fmt
         self._print_fps = print_fps
+
+        frame_shape = fmt.value[1](width, height)
+        if isinstance(frame_shape, int):
+            def check_frame_shape(frame: np.ndarray):
+                if frame.size != frame_shape:
+                    raise ValueError(f"unexpected frame size: {frame.size} != {frame_shape}")
+        else:
+            def check_frame_shape(frame: np.ndarray):
+                if frame.shape != frame_shape:
+                    raise ValueError(f"unexpected frame shape: {frame.shape} != {frame_shape}")
+
+        self._check_frame_shape = check_frame_shape
 
         self._fps_counter = FPSCounter(fps)
         self._fps_last_printed = time.perf_counter()
@@ -83,6 +111,10 @@ class Camera:
     @property
     def fps(self) -> float:
         return self._fps
+    
+    @property
+    def fmt(self) -> PixelFormat:
+        return self._fmt
 
     @property
     def frames_sent(self) -> int:
@@ -92,13 +124,10 @@ class Camera:
         self._backend.close()
 
     def send(self, frame: np.ndarray) -> None:
-        if frame.ndim != 3 or frame.shape[0] != self._height or frame.shape[1] != self._width:
-            raise ValueError(f"mismatching frame dimensions: {frame.shape} != ({self._height}, {self._width}, 3)")
-        if frame.shape[2] == 4:
-            warnings.warn('RGBA frames are deprecated, use RGB instead', DeprecationWarning)
-            frame = frame[:,:,:3]
-        elif frame.shape[2] != 3:
-            raise ValueError(f"invalid number of color channels, must be RGB or (deprecated) RGBA")
+        if frame.dtype != np.uint8:
+            raise TypeError(f'unexpected frame dtype: {frame.dtype} != uint8')
+        
+        self._check_frame_shape(frame)
 
         self._frames_sent += 1
         self._last_frame_t = time.perf_counter()