+++ pyvirtualcam/native_windows_obs/virtual_output.h
@@ -12,7 +12,8 @@ class VirtualOutput {
     video_queue_t *vq;
     uint32_t frame_width;
     uint32_t frame_height;
-    std::vector<uint8_t> buffer_i420;
+    uint32_t frame_fmt;
+    std::vector<uint8_t> buffer_tmp;
     std::vector<uint8_t> buffer_output;
     bool have_clockfreq = false;
     LARGE_INTEGER clock_freq;
@@ -34,7 +35,7 @@ class VirtualOutput {
     }
 
   public:
-    VirtualOutput(uint32_t width, uint32_t height, double fps) {
+    VirtualOutput(uint32_t width, uint32_t height, double fps, uint32_t fourcc) {
         // https://github.com/obsproject/obs-studio/blob/9da6fc67/.github/workflows/main.yml#L484
         LPCWSTR guid = L"CLSID\\{A3FCE0F5-3493-419F-958A-ABA1250EC20B}";
         HKEY key = nullptr;
@@ -45,6 +46,35 @@ class VirtualOutput {
             );
         }
 
+        frame_fmt = libyuv::CanonicalFourCC(fourcc);
+        frame_width = width;
+        frame_height = height;
+
+        uint32_t out_frame_size = nv12_frame_size(width, height);
+
+        switch (frame_fmt) {
+            case libyuv::FOURCC_RAW:
+            case libyuv::FOURCC_24BG:
+                // RGB|BGR -> I420 -> NV12
+                buffer_tmp.resize(i420_frame_size(width, height));
+                buffer_output.resize(out_frame_size);
+                break;
+            case libyuv::FOURCC_J400:
+                // GRAY -> ARGB -> NV12
+                buffer_tmp.resize(argb_frame_size(width, height));
+                buffer_output.resize(out_frame_size);
+                break;
+            case libyuv::FOURCC_I420:
+            case libyuv::FOURCC_YUY2:
+                // I420|YUYV -> NV12
+                buffer_output.resize(out_frame_size);
+                break;
+            default:
+                throw std::runtime_error(
+                    "Unsupported image format, must RGB, BGR, or I420."
+                );
+        }
+        
         uint64_t interval = (uint64_t)(10000000.0 / fps);
 
         vq = video_queue_create(width, height, interval);
@@ -53,10 +83,6 @@ class VirtualOutput {
             throw std::runtime_error("virtual camera output could not be started");
         }
 
-        frame_width = width;
-        frame_height = height;
-        buffer_i420.resize(i420_frame_size(width, height));
-        buffer_output.resize(nv12_frame_size(width, height));
         output_running = true;
     }
 
@@ -69,18 +95,36 @@ class VirtualOutput {
         output_running = false;
     }
 
-    // data is in RGB format (packed RGB, 24bpp, RGBRGB...)
-    // queue expects NV12 (semi-planar YUV, 12bpp)
-    void send(const uint8_t *rgb)
+    void send(const uint8_t *frame)
     {
         if (!output_running)
             return;
 
-        uint8_t* i420 = buffer_i420.data();
+        uint8_t* tmp = buffer_tmp.data();
         uint8_t* nv12 = buffer_output.data();
-
-        rgb_to_i420(rgb, i420, frame_width, frame_height);
-        i420_to_nv12(i420, nv12, frame_width, frame_height);
+        
+        switch (frame_fmt) {
+            case libyuv::FOURCC_RAW:
+                rgb_to_i420(frame, tmp, frame_width, frame_height);
+                i420_to_nv12(tmp, nv12, frame_width, frame_height);
+                break;
+            case libyuv::FOURCC_24BG:
+                bgr_to_i420(frame, tmp, frame_width, frame_height);
+                i420_to_nv12(tmp, nv12, frame_width, frame_height);
+                break;
+            case libyuv::FOURCC_J400:
+                gray_to_argb(frame, tmp, frame_width, frame_height);
+                argb_to_nv12(tmp, nv12, frame_width, frame_height);
+                break;
+            case libyuv::FOURCC_I420:
+                i420_to_nv12(frame, nv12, frame_width, frame_height);
+                break;
+            case libyuv::FOURCC_YUY2:
+                yuyv_to_nv12(frame, nv12, frame_width, frame_height);
+                break;
+            default:
+                throw std::logic_error("not implemented");
+        }
 
         // NV12 has two planes
         uint8_t* y = nv12;