+++ test/test_camera.py
@@ -1,9 +1,9 @@
 import os
 import platform
-import warnings
 import pytest
 import numpy as np
 import pyvirtualcam
+from pyvirtualcam import PixelFormat
 
 def test_consecutive():
     with pyvirtualcam.Camera(width=1280, height=720, fps=20) as cam:
@@ -51,12 +51,23 @@ def test_invalid_frame_shape():
         with pytest.raises(ValueError):
             cam.send(np.zeros((cam.height, cam.width), np.uint8))
 
-def test_deprecated_rgba_frame_format():
+def test_invalid_frame_dtype():
     with pyvirtualcam.Camera(width=1280, height=720, fps=20) as cam:
-        with warnings.catch_warnings():
-            warnings.simplefilter("error")
-            with pytest.raises(DeprecationWarning):
-                cam.send(np.zeros((cam.height, cam.width, 4), np.uint8))
+        with pytest.raises(TypeError):
+            cam.send(np.zeros((cam.height, cam.width, 3), np.uint16))
+
+def test_alternative_pixel_formats():
+    with pyvirtualcam.Camera(width=1280, height=720, fps=20, fmt=PixelFormat.BGR) as cam:
+        cam.send(np.zeros((cam.height, cam.width, 3), np.uint8))
+    
+    with pyvirtualcam.Camera(width=1280, height=720, fps=20, fmt=PixelFormat.GRAY) as cam:
+        cam.send(np.zeros((cam.height, cam.width), np.uint8))
+
+    with pyvirtualcam.Camera(width=1280, height=720, fps=20, fmt=PixelFormat.I420) as cam:
+        cam.send(np.zeros(cam.height * cam.width + cam.height * (cam.width // 2), np.uint8))
+    
+    with pyvirtualcam.Camera(width=1280, height=720, fps=20, fmt=PixelFormat.YUYV) as cam:
+        cam.send(np.zeros(cam.height * cam.width * 2, np.uint8))
 
 @pytest.mark.skipif(
     os.environ.get('CI') and platform.system() == 'Darwin',