+++ samples/video.py
@@ -1,4 +1,7 @@
 # This script plays back a video file on the virtual camera.
+# It also shows how to:
+# - select a specific camera device
+# - use BGR as pixel format
 
 import argparse
 import pyvirtualcam
@@ -8,6 +11,7 @@ import cv2
 parser = argparse.ArgumentParser()
 parser.add_argument("video_path", help="path to input video file")
 parser.add_argument("--fps", action="store_true", help="output fps every second")
+parser.add_argument("--device", help="virtual camera device, e.g. /dev/video0 (optional)")
 args = parser.parse_args()
 
 video = cv2.VideoCapture(args.video_path)
@@ -18,7 +22,8 @@ width = int(video.get(cv2.CAP_PROP_FRAME_WIDTH))
 height = int(video.get(cv2.CAP_PROP_FRAME_HEIGHT))
 fps = video.get(cv2.CAP_PROP_FPS)
 
-with pyvirtualcam.Camera(width, height, fps, fmt=PixelFormat.BGR, print_fps=args.fps) as cam:
+with pyvirtualcam.Camera(width, height, fps, fmt=PixelFormat.BGR,
+                         device=args.device, print_fps=args.fps) as cam:
     print(f'Virtual cam started: {cam.device} ({cam.width}x{cam.height} @ {cam.fps}fps)')
     count = 0
     while True: