+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Added
+- Support for common pixel formats: RGB (default), BGR (useful for OpenCV), GRAY, I420, YUYV.
+
+### Removed
+- RGBA support has been removed. Use RGB instead.
+
 ## [0.5.0] - 2012-03-13
 ### Added
 - Linux: multiple camera support (#37).